+++ bw.workflow.registercriminalclearancereport/src/main/java/sa/gov/nic/bio/bw/workflow/registercriminalclearancereport/controllers/ReviewAndSubmitPaneFxController.java
@@ -112,8 +112,7 @@ public class ReviewAndSubmitPaneFxController extends WizardStepFxControllerBase 
 
         if (normalizedPersonInfo != null) {
 
-
-            facePhotoBase64 = normalizedPersonInfo.getFacePhotoBase64();
+            if (facePhotoBase64 == null) { facePhotoBase64 = normalizedPersonInfo.getFacePhotoBase64(); }
 
             Name name = personInfo.getName();
             firstName = name.getFirstName();