+++ RELEASE-NOTES.md
@@ -9,6 +9,16 @@
 
 VERSION SCHEMA: vYYYY.MM.#SEQ
 
+## v2021.02.1
+
+- ENH: Disable recapturing fingerprints when register criminal clearance report by ID without role.
+- CHN: Enable to register unknown person when id start's with (0,8 or 9) when register convicted report.
+- FIX: Change SAUDI_ZONE from GMT+3 to Asia/Riyadh to fix issue with date before 1947.
+- CHN: Change iris registration service.
+- ENH: Updated biokit library to add iris score.
+
+---
+
 ## v2021.01.1
 
 - NEW: Register Criminal Clearance Report by ID.
@@ -44,7 +54,8 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 
 ## v2020.04.3
 
-- FIX: NPE in InquiryByFingerprintsResultPaneFxController.
+- FIX: 
+ in InquiryByFingerprintsResultPaneFxController.
 
 ---